Module Parser

This module provides abstract interfacing for parsing input files.

Note: Each new file format added to onijn is a concrete implementation of this abstract interfacing.

module type PARSER = sig ... end

This module type provides an abstract interface for parsing.

module ParseWithErrors (P : PARSER) : sig ... end

This functor provides a default implementation for a concrete parser that support 'exit on error' parsing.