Module Syntax.Symb

This module exports a module type NAME, which is an abstract interface for names, and one module IndexedName of type NAME.

module type NAME = sig ... end

Interface for names. It can be used to represent base types, variables, and function symbols.

module IndexedName () : NAME